disease_model/requirements.txt
Flask
flask-cors
python-dotenv
gunicorn
pandas
numpy
scikit-learn
catboost
